requirements.txt
bcrypt==4.2.1
blinker==1.9.0
cachelib==0.13.0
click==8.1.7
dnspython==2.7.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-PyMongo==2.3.0
flask-redis==0.4.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgspec==0.18.6
psycopg2==2.9.10
PyJWT==2.10.0
pymongo==4.10.1
python-dotenv==1.0.1
redis==5.2.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
